Struct starchart_backends::fs::FsError

pub struct FsError { /* private fields */ }

This is supported on crate feature fs only.

An error occurred from the FsBackend or one of it’s Transcoders.

Implementations

impl FsError

pub fn serde(err: Option<Box<dyn Error + Send + Sync>>) -> Self

Creates an error from a Transcoder.

pub const fn kind(&self) -> &FsErrorType

Immutable reference to the type of error that occurred.

pub fn into_source(self) -> Option<Box<dyn Error + Send + Sync>>

Consume the error, returning the source error if there is any.

pub fn into_parts(self) -> (FsErrorType, Option<Box<dyn Error + Send + Sync>>)

Consume the error, returning the owned error type and the source error.